Simple multithreaded actor critic
Implements threading in the LOLAgent model.

Open Questions:
1. Threading sanity. Currently implements threading using the threading
library. So there's no interleaving in the python code due to GIL.
However, there's plenty of opportunities for parallelism in tf and np
code.

2. Shared optimization. The tf RMSProp optimizer should keep moving
average values for each parameter for each thread. The moving averages
should be shared.

3. The current weight update scheme is hacky. The gradients should be
applied on a fresh copy of the global parameters. Now the global
parameters are computed as a weighted average of the current global
parameters and the thread specific parameters.