[WIP] Adds LSTM Cells after the hidden layer
A3C paper features a variant of the model with LSTM cells after the last
hidden layer in the network. This commit adds LSTM cells to the network.

Working: Network initializes and algorithm runs.
Not working: The network doesn't seem to learn.

TODO:
Check network layout.
Change from python deques to np.ndarray based ringbuffers in the
training loop to avoid back-and-forth conversions.