[WIP] Refactor RNN logic
Makes the RNN logic executable.

The RNN layer takes as inputs the last hidden states of the feedforward
network evaluated on consecutive observations. The RNN outputs are fed
to the softmax and linear layers for policy and value computation
respectively. The final state of the RNN is passed to the next
evaluation but zeroed out in the beginning of each episode.

Could not find suitable parameters to solve CartPole in initial
exploration. There might be some problems left in the rnn
implementation.

Makes the FF and RNN codepaths more compatible so they can be switched
with a flag.

TODO:
Check RNN implementation logic.
Try bucketing instead of dynamic_rnn for performance.
Change to multicell rnn.